Header files were reorgnized by using pre-complied header.

+++ G2D_cpp/setupRainfall.cpp
@@ -1,15 +1,7 @@
 #include "stdafx.h"
-#include <stdio.h>
-#include <fstream>
-#include <filesystem>
-#include <io.h>
-#include<ATLComTime.h>
-#include <string>
-#include <omp.h>
 #include "g2d.h"
 #include "gentle.h"
 
-//#include "g2d_cuda.cuh"
 
 using namespace std;
 
@@ -100,14 +92,14 @@ int readRainfallAndGetIntensity(int rforder)
 	if ((rforder - 1) < (int)rf.size())//강우자료 있으면, 읽어서 세팅
 	{
 		double inRF_mm = 0;
-		int rfIntervalSEC = prj.rainfallDataInterval_min * 60;
+		double rfIntervalSEC = prj.rainfallDataInterval_min * 60.0;
 		rainfallDataType rftype = prj.rainfallDataType;
 		switch (rftype)
 		{
 		case rainfallDataType::TextFileMAP:
 			inRF_mm = stof(rf[rforder - 1].rainfall);
 			if (inRF_mm < 0) { inRF_mm = 0.0; }
-			psi.rfReadintensityForMAP_mPsec = inRF_mm / 1000.0f / rfIntervalSEC;
+			psi.rfReadintensityForMAP_mPsec = inRF_mm / 1000.0 / rfIntervalSEC;
 			// 우선 여기에 저장했다가, cvs 초기화 할때 셀별로 배분한다. 시간 단축을 위해서
 			break;
 		case rainfallDataType::TextFileASCgrid:
